+++ xarray/tests/test_backends.py
@@ -2261,7 +2261,8 @@ def test_open_mfdataset_manyfiles(readengine, nfiles, parallel, chunks,
             subds.to_netcdf(tmpfiles[ii], engine=writeengine)
 
         # check that calculation on opened datasets works properly
-        with open_mfdataset(tmpfiles, engine=readengine, parallel=parallel,
+        with open_mfdataset(tmpfiles, combine='manual', concat_dim='x',
+                            engine=readengine, parallel=parallel,
                             chunks=chunks) as actual:
 
             # check that using open_mfdataset returns dask arrays for variables
@@ -2317,11 +2318,13 @@ class TestOpenMFDatasetWithDataVarsAndCoordsKw:
     @pytest.mark.parametrize('opt', ['all', 'minimal', 'different'])
     def test_open_mfdataset_does_same_as_concat(self, opt):
         with self.setup_files_and_datasets() as (files, [ds1, ds2]):
-            with open_mfdataset(files, data_vars=opt) as ds:
+            with open_mfdataset(files, data_vars=opt,
+                                combine='manual', concat_dim='t') as ds:
                 kwargs = dict(data_vars=opt, dim='t')
                 ds_expect = xr.concat([ds1, ds2], **kwargs)
                 assert_identical(ds, ds_expect)
-            with open_mfdataset(files, coords=opt) as ds:
+            with open_mfdataset(files, coords=opt,
+                                combine='manual', concat_dim='t') as ds:
                 kwargs = dict(coords=opt, dim='t')
                 ds_expect = xr.concat([ds1, ds2], **kwargs)
                 assert_identical(ds, ds_expect)
@@ -2331,7 +2334,8 @@ class TestOpenMFDatasetWithDataVarsAndCoordsKw:
 
         with self.setup_files_and_datasets() as (files, [ds1, ds2]):
             # open the files with the data_var option
-            with open_mfdataset(files, data_vars=opt) as ds:
+            with open_mfdataset(files, data_vars=opt,
+                                combine='manual', concat_dim='t') as ds:
 
                 coord_shape = ds[self.coord_name].shape
                 coord_shape1 = ds1[self.coord_name].shape
@@ -2348,7 +2352,8 @@ class TestOpenMFDatasetWithDataVarsAndCoordsKw:
 
         with self.setup_files_and_datasets() as (files, [ds1, ds2]):
             # open the files using data_vars option
-            with open_mfdataset(files, data_vars=opt) as ds:
+            with open_mfdataset(files, data_vars=opt,
+                                combine='manual', concat_dim='t') as ds:
 
                 coord_shape = ds[self.coord_name].shape
                 coord_shape1 = ds1[self.coord_name].shape
@@ -2364,12 +2369,14 @@ class TestOpenMFDatasetWithDataVarsAndCoordsKw:
 
         with self.setup_files_and_datasets() as (files, _):
             with pytest.raises(ValueError):
-                with open_mfdataset(files, data_vars='minimum'):
+                with open_mfdataset(files, data_vars='minimum',
+                                    combine='auto'):
                     pass
 
             # test invalid coord parameter
             with pytest.raises(ValueError):
-                with open_mfdataset(files, coords='minimum'):
+                with open_mfdataset(files, coords='minimum',
+                                    combine='auto'):
                     pass
 
 
@@ -2437,11 +2444,14 @@ class TestDask(DatasetIOBase):
             with create_tmp_file() as tmp2:
                 original.isel(x=slice(5)).to_netcdf(tmp1)
                 original.isel(x=slice(5, 10)).to_netcdf(tmp2)
-                with open_mfdataset([tmp1, tmp2]) as actual:
+                with open_mfdataset([tmp1, tmp2], concat_dim='x',
+                                    combine='manual') as actual:
                     assert isinstance(actual.foo.variable.data, da.Array)
                     assert actual.foo.variable.data.chunks == ((5, 5),)
                     assert_identical(original, actual)
-                with open_mfdataset([tmp1, tmp2], chunks={'x': 3}) as actual:
+                with open_mfdataset([tmp1, tmp2], concat_dim='x',
+                                    combine='manual',
+                                    chunks={'x': 3}) as actual:
                     assert actual.foo.variable.data.chunks == ((3, 2, 3, 2),)
 
         with raises_regex(IOError, 'no files to open'):
@@ -2466,6 +2476,7 @@ class TestDask(DatasetIOBase):
                                       y=slice(4, 8)).to_netcdf(tmp4)
                         with open_mfdataset([[tmp1, tmp2],
                                              [tmp3, tmp4]],
+                                            combine='manual',
                                             concat_dim=['y', 'x']) as actual:
                             assert isinstance(actual.foo.variable.data,
                                               da.Array)
@@ -2474,6 +2485,7 @@ class TestDask(DatasetIOBase):
                             assert_identical(original, actual)
                         with open_mfdataset([[tmp1, tmp2],
                                              [tmp3, tmp4]],
+                                            combine='manual',
                                             concat_dim=['y', 'x'],
                                             chunks={'x': 3, 'y': 2}) as actual:
                             assert actual.foo.variable.data.chunks == \
@@ -2488,7 +2500,8 @@ class TestDask(DatasetIOBase):
                 tmp2 = Path(tmp2)
                 original.isel(x=slice(5)).to_netcdf(tmp1)
                 original.isel(x=slice(5, 10)).to_netcdf(tmp2)
-                with open_mfdataset([tmp1, tmp2]) as actual:
+                with open_mfdataset([tmp1, tmp2], concat_dim='x',
+                                    combine='manual') as actual:
                     assert_identical(original, actual)
 
     @requires_pathlib
@@ -2512,10 +2525,10 @@ class TestDask(DatasetIOBase):
                                       y=slice(4, 8)).to_netcdf(tmp4)
                         with open_mfdataset([[tmp1, tmp2],
                                              [tmp3, tmp4]],
+                                            combine='manual',
                                             concat_dim=['y', 'x']) as actual:
                             assert_identical(original, actual)
 
-    @pytest.mark.xfail(reason="Not yet implemented")
     def test_open_mfdataset_2(self):
         original = Dataset({'foo': ('x', np.random.randn(10))})
         with create_tmp_file() as tmp1:
@@ -2523,11 +2536,8 @@ class TestDask(DatasetIOBase):
                 original.isel(x=slice(5)).to_netcdf(tmp1)
                 original.isel(x=slice(5, 10)).to_netcdf(tmp2)
 
-                with pytest.raises(NotImplementedError):
-                    open_mfdataset([tmp1, tmp2], infer_order_from_coords=True)
-
-                # With infer_order_from_coords=True this should pass in future
-                with open_mfdataset([tmp1, tmp2]) as actual:
+                with open_mfdataset([tmp1, tmp2], concat_dim='x',
+                                    combine='manual') as actual:
                     assert_identical(original, actual)
 
     def test_attrs_mfdataset(self):
@@ -2540,7 +2550,8 @@ class TestDask(DatasetIOBase):
                 ds2.attrs['test2'] = 'bar'
                 ds1.to_netcdf(tmp1)
                 ds2.to_netcdf(tmp2)
-                with open_mfdataset([tmp1, tmp2]) as actual:
+                with open_mfdataset([tmp1, tmp2], concat_dim='x',
+                                    combine='manual') as actual:
                     # presumes that attributes inherited from
                     # first dataset loaded
                     assert actual.test1 == ds1.test1
@@ -2548,6 +2559,49 @@ class TestDask(DatasetIOBase):
                     with raises_regex(AttributeError,
                                       'no attribute'):
                         actual.test2
+
+    def test_open_mfdataset_auto_combine(self):
+        original = Dataset({'foo': ('x', np.random.randn(10)),
+                            'x': np.arange(10)})
+        with create_tmp_file() as tmp1:
+            with create_tmp_file() as tmp2:
+                original.isel(x=slice(5)).to_netcdf(tmp1)
+                original.isel(x=slice(5, 10)).to_netcdf(tmp2)
+
+                with open_mfdataset([tmp2, tmp1], combine='auto') as actual:
+                    assert_identical(original, actual)
+
+    def test_open_mfdataset_combine_manual_no_concat_dim(self):
+        original = Dataset({'foo': ('x', np.random.randn(10)),
+                            'x': np.arange(10)})
+        with create_tmp_file() as tmp1:
+            with create_tmp_file() as tmp2:
+                original.isel(x=slice(5)).to_netcdf(tmp1)
+                original.isel(x=slice(5, 10)).to_netcdf(tmp2)
+
+                with raises_regex(ValueError, 'Must supply concat_dim'):
+                    open_mfdataset([tmp2, tmp1], combine='manual')
+
+    @pytest.mark.xfail(reason='mfdataset loses encoding currently.')
+    def test_encoding_mfdataset(self):
+        original = Dataset({'foo': ('t', np.random.randn(10)),
+                            't': ('t', pd.date_range(start='2010-01-01',
+                                                     periods=10,
+                                                     freq='1D'))})
+        original.t.encoding['units'] = 'days since 2010-01-01'
+
+        with create_tmp_file() as tmp1:
+            with create_tmp_file() as tmp2:
+                ds1 = original.isel(t=slice(5))
+                ds2 = original.isel(t=slice(5, 10))
+                ds1.t.encoding['units'] = 'days since 2010-01-01'
+                ds2.t.encoding['units'] = 'days since 2000-01-01'
+                ds1.to_netcdf(tmp1)
+                ds2.to_netcdf(tmp2)
+                with open_mfdataset([tmp1, tmp2]) as actual:
+                    assert actual.t.encoding['units'] == original.t.encoding['units']  # noqa
+                    assert actual.t.encoding['units'] == ds1.t.encoding['units']  # noqa
+                    assert actual.t.encoding['units'] != ds2.t.encoding['units']  # noqa
 
     def test_preprocess_mfdataset(self):
         original = Dataset({'foo': ('x', np.random.randn(10))})
@@ -2558,7 +2612,8 @@ class TestDask(DatasetIOBase):
                 return ds.assign_coords(z=0)
 
             expected = preprocess(original)
-            with open_mfdataset(tmp, preprocess=preprocess) as actual:
+            with open_mfdataset(tmp, preprocess=preprocess,
+                                combine='auto') as actual:
                 assert_identical(expected, actual)
 
     def test_save_mfdataset_roundtrip(self):
@@ -2568,7 +2623,8 @@ class TestDask(DatasetIOBase):
         with create_tmp_file() as tmp1:
             with create_tmp_file() as tmp2:
                 save_mfdataset(datasets, [tmp1, tmp2])
-                with open_mfdataset([tmp1, tmp2]) as actual:
+                with open_mfdataset([tmp1, tmp2], concat_dim='x',
+                                    combine='manual') as actual:
                     assert_identical(actual, original)
 
     def test_save_mfdataset_invalid(self):
@@ -2594,14 +2650,15 @@ class TestDask(DatasetIOBase):
                 tmp1 = Path(tmp1)
                 tmp2 = Path(tmp2)
                 save_mfdataset(datasets, [tmp1, tmp2])
-                with open_mfdataset([tmp1, tmp2]) as actual:
+                with open_mfdataset([tmp1, tmp2], concat_dim='x',
+                                    combine='manual') as actual:
                     assert_identical(actual, original)
 
     def test_open_and_do_math(self):
         original = Dataset({'foo': ('x', np.random.randn(10))})
         with create_tmp_file() as tmp:
             original.to_netcdf(tmp)
-            with open_mfdataset(tmp) as ds:
+            with open_mfdataset(tmp, combine='auto') as ds:
                 actual = 1.0 * ds
                 assert_allclose(original, actual, decode_bytes=False)
 
@@ -2611,7 +2668,8 @@ class TestDask(DatasetIOBase):
                 data = Dataset({'x': 0})
                 data.to_netcdf(tmp1)
                 Dataset({'x': np.nan}).to_netcdf(tmp2)
-                with open_mfdataset([tmp1, tmp2], concat_dim=None) as actual:
+                with open_mfdataset([tmp1, tmp2], concat_dim=None,
+                                    combine='manual') as actual:
                     assert_identical(data, actual)
 
     def test_open_dataset(self):
@@ -2638,7 +2696,8 @@ class TestDask(DatasetIOBase):
                            {'baz': [100]})
         with create_tmp_file() as tmp:
             original.to_netcdf(tmp)
-            with open_mfdataset([tmp], concat_dim=dim) as actual:
+            with open_mfdataset([tmp], concat_dim=dim,
+                                combine='manual') as actual:
                 assert_identical(expected, actual)
 
     def test_open_multi_dataset(self):
@@ -2661,7 +2720,8 @@ class TestDask(DatasetIOBase):
                 create_tmp_file() as tmp2:
             original.to_netcdf(tmp1)
             original.to_netcdf(tmp2)
-            with open_mfdataset([tmp1, tmp2], concat_dim=dim) as actual:
+            with open_mfdataset([tmp1, tmp2], concat_dim=dim,
+                                combine='manual') as actual:
                 assert_identical(expected, actual)
 
     def test_dask_roundtrip(self):
@@ -2680,10 +2740,10 @@ class TestDask(DatasetIOBase):
         with create_tmp_file() as tmp:
             data = create_test_data()
             data.to_netcdf(tmp)
-            with open_mfdataset(tmp) as ds:
+            with open_mfdataset(tmp, combine='auto') as ds:
                 original_names = dict((k, v.data.name)
                                       for k, v in ds.data_vars.items())
-            with open_mfdataset(tmp) as ds:
+            with open_mfdataset(tmp, combine='auto') as ds:
                 repeat_names = dict((k, v.data.name)
                                     for k, v in ds.data_vars.items())
             for var_name, dask_name in original_names.items():
@@ -2713,7 +2773,8 @@ class TestDask(DatasetIOBase):
                                              engine=self.engine, compute=False)
                 assert isinstance(delayed_obj, Delayed)
                 delayed_obj.compute()
-                with open_mfdataset([tmp1, tmp2]) as actual:
+                with open_mfdataset([tmp1, tmp2], combine='manual',
+                                    concat_dim='x') as actual:
                     assert_identical(actual, original)
 
     def test_load_dataset(self):
@@ -2732,6 +2793,59 @@ class TestDask(DatasetIOBase):
             # this would fail if we used open_dataarray instead of
             # load_dataarray
             ds.to_netcdf(tmp)
+
+
+@requires_scipy_or_netCDF4
+class TestOpenMFDataSetDeprecation:
+    """
+    Set of tests to check that FutureWarnings are correctly raised until the
+    deprecation cycle is complete. #2616
+    """
+    def test_open_mfdataset_with_concat_dim(self):
+        ds1, ds2 = Dataset({'x': [0]}), Dataset({'x': [1]})
+        with create_tmp_file() as tmp1:
+            with create_tmp_file() as tmp2:
+                ds1.to_netcdf(tmp1)
+                ds2.to_netcdf(tmp2)
+
+                with pytest.warns(FutureWarning, match="`concat_dim`"):
+                    open_mfdataset([tmp1, tmp2], concat_dim='x')
+
+    def test_auto_combine_with_merge_and_concat(self):
+        ds1, ds2 = Dataset({'x': [0]}), Dataset({'x': [1]})
+        ds3 = Dataset({'z': ((), 99)})
+        with create_tmp_file() as tmp1:
+            with create_tmp_file() as tmp2:
+                with create_tmp_file() as tmp3:
+                    ds1.to_netcdf(tmp1)
+                    ds2.to_netcdf(tmp2)
+                    ds3.to_netcdf(tmp3)
+
+                    with pytest.warns(FutureWarning,
+                                      match="require both concatenation"):
+                        open_mfdataset([tmp1, tmp2, tmp3])
+
+    def test_auto_combine_with_coords(self):
+        ds1 = Dataset({'foo': ('x', [0])}, coords={'x': ('x', [0])})
+        ds2 = Dataset({'foo': ('x', [1])}, coords={'x': ('x', [1])})
+        with create_tmp_file() as tmp1:
+            with create_tmp_file() as tmp2:
+                ds1.to_netcdf(tmp1)
+                ds2.to_netcdf(tmp2)
+
+                with pytest.warns(FutureWarning, match="supplied have global"):
+                    open_mfdataset([tmp1, tmp2])
+
+    def test_auto_combine_without_coords(self):
+        ds1, ds2 = Dataset({'foo': ('x', [0])}), Dataset({'foo': ('x', [1])})
+        with create_tmp_file() as tmp1:
+            with create_tmp_file() as tmp2:
+                ds1.to_netcdf(tmp1)
+                ds2.to_netcdf(tmp2)
+
+                with pytest.warns(FutureWarning,
+                                  match="supplied do not have global"):
+                    open_mfdataset([tmp1, tmp2])
 
 
 @requires_scipy_or_netCDF4
@@ -3052,7 +3166,7 @@ class TestPseudoNetCDFFormat:
             ['example.uamiv',
              'example.uamiv'],
             engine='pseudonetcdf',
-            concat_dim=['TSTEP'],
+            concat_dim='TSTEP',
             backend_kwargs={'format': 'uamiv'})
 
         data1 = np.arange(20, dtype='f').reshape(1, 1, 4, 5)